requirements.txt
pandas==0.24.2
Twisted==19.7.0
psycopg2-binary==2.8.4
scipy==1.3.0
Scrapy==1.7.4
whois==0.9.3
lime==0.1.1.37
shap==0.38.1
prettytable==0.7.2
seaborn==0.9.0
matplotlib==3.1.0
lxml==4.4.1
requests==2.22.0
xgboost==0.90
numpy==1.16.4
beautifulsoup4==4.8.2
scikit_learn==0.24.1
selenium==3.141.0
requests==2.22.0
url-normalize==1.4.2
setuptools>=46.0.0
kosoh_logo_classifier==0.1
